ProgressButtonを別ファイルに移動
